chore: organize the `blackbox_solver` crate
