+++ acvm-repo/blackbox_solver/src/ecdsa/mod.rs
@@ -1,0 +1,22 @@
+use crate::BlackBoxResolutionError;
+
+mod secp256r1;
+mod secp256k1;
+
+pub fn ecdsa_secp256k1_verify(
+    hashed_msg: &[u8],
+    public_key_x: &[u8; 32],
+    public_key_y: &[u8; 32],
+    signature: &[u8; 64],
+) -> Result<bool, BlackBoxResolutionError> {
+    Ok(secp256k1::verify_signature(hashed_msg, public_key_x, public_key_y, signature))
+}
+
+pub fn ecdsa_secp256r1_verify(
+    hashed_msg: &[u8],
+    public_key_x: &[u8; 32],
+    public_key_y: &[u8; 32],
+    signature: &[u8; 64],
+) -> Result<bool, BlackBoxResolutionError> {
+    Ok(secp256r1::verify_signature(hashed_msg, public_key_x, public_key_y, signature))
+}
